No longer code splitting at the routing level, cleaned Home folder structure
