+++ src/pages/home/Home.jsx
@@ -1,21 +1,21 @@
+import React, { Suspense } from 'react';
 import { useOutletContext } from 'react-router';
 import { Helmet } from 'react-helmet';
-import { Link } from 'react-router-dom';
-import { useMediaQuery } from 'react-responsive';
 import AppearingDiv from '../../components/ui/AppearingDiv/AppearingDiv';
-import HomeTitle from './Components/HomeTitle';
-import Testimonies from './Components/HomeTestimonies';
-import HomeSocialsAndSignup from './Components/HomeSocialsAndSignup/HomeSocialsAndSignup';
+import HomeTitle from './Components/HomeTitle/HomeTitle';
 import useWindowDimensions from '../../hooks/useWindowDimensions';
-import useScrollPosition from './../../hooks/useScrollPosition';
+
+// Async Components for code splitting
+const HomeIntroduction = React.lazy(() => import('./Components/HomeIntroduction/HomeIntroduction'));
+const HomeLinkGallery = React.lazy(() => import('./Components/HomeLinkGallery/HomeLinkGallery'));
+const HomeSocialsAndSignup = React.lazy(() => import('./Components/HomeSocialsAndSignup/HomeSocialsAndSignup'));
+const Testimonies = React.lazy(() => import('./Components/HomeTestimonies/HomeTestimonies'));
 
 import './Home.css';
 
 export default function Home () {
     const {setNavbarScrollPosition, setNavbarScrollPositionMobile, setHide} = useOutletContext();
     const {height} = useWindowDimensions(); 
-    const scrollPosition = useScrollPosition()
-    const isMobile = useMediaQuery({ maxWidth: '1000px' });
 
     setNavbarScrollPosition(height - 70);
     setNavbarScrollPositionMobile(430);
@@ -25,61 +25,17 @@ export default function Home () {
         <>
             <Helmet><title>Home</title></Helmet>
             <HomeTitle/>
-            <div className='hello_container' id='helloyfa'>
-                <div className="scrolling_text" style={{ transform: `translateX(-${scrollPosition/3 + 1000}px)`, top: '20px'}}>
-                    THE ORGANIZATION OF YOUNG FILIPINO AMERICANS • THE ORGANIZATION OF YOUNG FILIPINO AMERICANS • THE ORGANIZATION OF YOUNG FILIPINO AMERICANS
-                </div>    
-                <AppearingDiv className='hello_text_container' translateMeasurement={50}>
-                    <h1>Helloyfa!</h1>
-                    <p>
-                        Since 1988, OYFA has been the University of Virginia's exclusive Filipino-American undergraduate CIO.
-                        OYFA exists to engage all in cultural understanding through its yearly live Barrio performance,
-                        positively impact the local community through acts of service and fundraising, and secure a space 
-                        for all those who are any part Young, Filipino, and American, or not at all.
-                    </p>
-                    <Link to='/about'>About Us</Link>
-                </AppearingDiv>
-                <div className="scrolling_text" style={{ transform: `translateX(${scrollPosition/3 - 1400}px)`, bottom: '20px'}}>
-                    THE ORGANIZATION OF YOUNG FILIPINO AMERICANS • THE ORGANIZATION OF YOUNG FILIPINO AMERICANS • THE ORGANIZATION OF YOUNG FILIPINO AMERICANS • THE ORGANIZATION OF YOUNG FILIPINO AMERICANS
+
+            <Suspense fallback={<div>Loading...</div>}>
+                <HomeIntroduction/>
+                <HomeLinkGallery/>
+                <HomeSocialsAndSignup/>
+                <div className='other_container'>
+                    <img src='images/home/ffn.jpg'/>
+                    <AppearingDiv className='thank_you'><h3>Thank you, 36th &lt;3</h3></AppearingDiv>
                 </div>
-                <AppearingDiv className='hello_photo_box2'/>
-                <AppearingDiv className='hello_photo' translateAxis='X' translateMeasurement={250} delay={0.7}>
-                    <img src='/images/home/Intro_Culturefest_Panorama.jpeg'/>
-                </AppearingDiv>
-            </div>
-
-            <div className='link_gallery'>
-                <Link to='/events' className='link_gallery_column'>
-                    <div className='link_gallery_photo_overlay'/>
-                    <AppearingDiv delay={0.7} margin={isMobile ? '0px 100px -100px 0px' : '0px 100px -40% 0px'} className='link_gallery_title'>
-                        <h1 className='link_gallery_top_title'>Events</h1>
-                    </AppearingDiv>
-                    <img className='link_gallery_photo' src='/images/home/Events.jpg'/>
-                </Link>
-                <Link to='/leadership' className='link_gallery_column'>
-                    <div className='link_gallery_photo_overlay'/>
-                    <AppearingDiv delay={0.9} margin={isMobile ? '0px 100px -100px 0px' : '0px 100px 0px 0px'} className='link_gallery_title'>
-                        <h1 className='link_gallery_bottom_title'>Leadership</h1>
-                        </AppearingDiv>
-                    <img className='link_gallery_photo_middle' src='/images/home/Leadership.jpg'/>
-                </Link>
-                <a href='https://uvaoyfa.myportfolio.com/oyfa-2023-2024' className='link_gallery_column'>
-                    <div className='link_gallery_photo_overlay'/>
-                    <AppearingDiv delay={1.1} margin={isMobile ? '0px 100px -100px 0px' : '0px 100px -40% 0px'} className='link_gallery_title'>
-                        <h1 className='link_gallery_top_title'>Photos</h1>
-                    </AppearingDiv>
-                    <img className='link_gallery_photo' src='/images/home/Photos.jpg'/>
-                </a>
-            </div>
-
-            <HomeSocialsAndSignup/>
-
-            <div className='other_container'>
-                <img src='images/home/ffn.jpg'/>
-                <AppearingDiv className='thank_you'><h3>Thank you, 36th &lt;3</h3></AppearingDiv>
-            </div>
-        
-            <Testimonies/>
+                <Testimonies/>
+            </Suspense>
         </>
     )
 }
