--- src/pages/home/Stylesheets/HomeIntro.css
@@ -1,123 +1,0 @@
-.intro_parent{
-    position: relative;
-    width: 100%;
-    height: 80%;
-    /* background-image: linear-gradient(#aec6ff 0%, var(--off-white) 25%); */
-    /*background-image: linear-gradient(to right, #ff5569 0%, var(--off-white) 20%, var(--off-white) 80%, #ffffb8 100%);*/
-    /*border-image: linear-gradient(to right, #CE1126 0%, #0038A8 33%, #051941 66%, #FFFF00 100%) 1;*/
-    /*(alternate) border-image: linear-gradient(to right, #CE1126 0%, #0038A8 33%, #FFFF00 67%, #FFFFFF 100%) 1;*/
-
-    border-image: linear-gradient(to right, #841F27 0%, #336699 33%, #0c2e73 66%, #ffb300 100%) 1;
-    border-width: 1vh;
-    border-style: solid;
-    
-    
-    padding-bottom: 5%;
-}
-
-.intro_paragraph_container{
-    width: 80%;
-    display: flex;
-    margin: 0 auto;
-    text-align: center;
-}
-
-.intro_social_media_parent{
-    width: 50%;
-    display: flex;
-    margin: 0 auto;
-}
-
-.intro_heading{
-    font-family: 'Poppins';
-    /*font-size: 8em;*/
-    font-size: 15vh;
-    font-weight: bold;
-    padding-top: 3%;
-    margin-bottom: 3%;
-    text-align: center;
-}
-
-.intro_get_involved_heading{
-    font-family: 'Poppins';
-    /*font-size: 6em;*/
-    font-size: 14vh;
-    font-weight: bold;
-    padding-top: 3%;
-    margin-bottom: 3%;
-    text-align: center;
-}
-
-.intro_paragraph{
-    font-family: 'Open Sans';
-    font-size: 1.5em;
-    margin-top: 3%;
-}
-
-.intro_links_container{
-    display: flex;
-    flex-direction: column;
-    margin: 0 auto;
-    text-align: center;
-    margin-top: 3%;
-}
-
-.intro_link{
-    font-family: 'Open Sans';
-    font-size: 1.5em;
-    margin-bottom: 3%;
-}
-
-.intro_react_link{
-    text-decoration: none;
-    color:black;
-}
-
-.intro_caret{
-    margin-left: 1%;
-    font-size: 1.2em;
-}
-
-.intro_img{
-    display: flex;
-    margin: 0 auto;
-    width: 80%;
-    height: 50vh;
-    object-fit: cover;
-    box-shadow: 0.5vh 0.5vh 0.5vh rgb(141, 141, 141);
-}
-
-@media screen and (max-width : 1000px) /* Phone Rules */
-{
-    /* decrease the size of stuff, throw the links into a flexbox to limit width and auto-c */
-    .intro_parent{
-        border-width: 0vh;
-    }
-    .intro_heading{
-        font-size: 8vh;
-        padding-top: 8%;
-        margin-bottom: 8%;
-    }
-    .intro_paragraph{
-        font-size: 2.5vh;
-    }
-    .intro_links_container{
-        max-width: 80%;
-    }
-    .intro_link{
-        font-size: 2.5vh;
-    }
-    .intro_caret{
-        font-size: 2.5vh;
-    }
-    .intro_img{
-        width: 100%;
-        height: 35vh;
-        box-shadow: none;
-    }
-    .intro_get_involved_heading{
-        font-size: 6vh;
-        padding-top: 8%;
-        margin-bottom: 8%;
-    }
-}
